리팩토링: 스타일 분리를 위해 프로젝트 구조 변경 - 1
- 컴포넌트 내에서 스타일을 분리하여 관리하고, 더 명시적인 스타일명을 사용하기 위해 구조 변경
