+++ src/components/Header/style.js
@@ -1,0 +1,23 @@
+import styled from '@emotion/styled';
+
+export const Div = {
+  Container: styled.div({
+    display: 'grid',
+    gridTemplateColumns: 'auto 300px',
+    backgroundColor: '#5383e8',
+    padding: '20px 40px 20px 40px',
+    alignItems: 'center',
+    a: {
+      fontSize: '1.8em',
+      fontWeight: 'bold',
+      display: 'block',
+      margin: '1em 0',
+      float: 'left',
+      textAlign: 'center',
+      color: '#FFF',
+      textDecoration: 'none',
+    },
+  }),
+};
+
+export const temp = '';
