+++ src/components/search/SearchBox/SearchBox.jsx
@@ -1,13 +1,5 @@
 import { memo } from 'react';
-import styled from '@emotion/styled';
-
-const SearchContainer = styled.div({
-  display: 'grid',
-  gridTemplateColumns: 'auto 50px',
-  position: 'relative',
-  height: '30px',
-  borderRadius: '4px',
-});
+import * as Styled from './style';
 
 function SearchBox({ userName, onChangeValue, onClickSearchPlayer }) {
   const handleChange = (event) => {
@@ -16,7 +8,7 @@ function SearchBox({ userName, onChangeValue, onClickSearchPlayer }) {
   };
 
   return (
-    <SearchContainer>
+    <Styled.Div.Container>
       <input
         type="text"
         placeholder="소환사명을 입력하세요."
@@ -29,7 +21,7 @@ function SearchBox({ userName, onChangeValue, onClickSearchPlayer }) {
       >
         검색
       </button>
-    </SearchContainer>
+    </Styled.Div.Container>
   );
 }
