+++ src/components/search/SearchBox/SearchBox.test.jsx
