+++ src/components/search/SearchBox/style.js
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const Div = {
+  Container: styled.div({
+    display: 'grid',
+    gridTemplateColumns: 'auto 50px',
+    position: 'relative',
+    height: '30px',
+    borderRadius: '4px',
+  }),
+};
+
+export const temp = '';
